03 09 21 Update

+++ app/Http/Controllers/Frontend/Friendship/GroupController.php
@@ -7,7 +7,9 @@ use App\Models\Access\User\User;
 use App\Models\Friendships\Group;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
+use App\Http\Requests\Backend\Groups\CreateGroupsRequest;
 use App\Models\Friendships\FriendFriendshipGroups;
+
 
 class GroupController extends Controller
 {
@@ -39,7 +41,7 @@ class GroupController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CreateGroupsRequest $request)
     {
         $group = new Group();
         $group->name = $request->group_name;
@@ -105,17 +107,57 @@ class GroupController extends Controller
 
     public function storeGroupFriends(Request $request)
     {
-        // dd($request->all());
-        $group = Group::find($request->group_id);
-        $friends = explode(',', $request->friends);
-        FriendFriendshipGroups::where('group_id', $group->id)->delete();
-        
-        for($i = 0; $i < count($friends); $i++) {
-            $friend = User::where('id',$friends[$i])->first();
-            // dd($group_name,$friend);
-            Auth::user()->groupFriend($friend, $group->id);
+        // dd($request->group_id);
+        if($request->group_id!='')
+        {
+            $group = Group::find($request->group_id);
+            // dd($group);
+            if($group)
+            {
+                $friends = explode(',', $request->friends);
+                
+                if(count($friends) > 1)
+                {
+                    FriendFriendshipGroups::where('group_id', $group->id)->delete();
+                    for($i = 0; $i < count($friends); $i++) {
+                        $friend = User::where('id',$friends[$i])->first();
+                        // dd($group_name,$friend);
+                        Auth::user()->groupFriend($friend, $group->id);
+                    }
+                    return array(
+                        'status'=>'success',
+                        'title'=>'Group Update!',
+                        'message' => 'Members of the group '.$group->name.' succussfully updated '
+                    );
+                }
+                else
+                {
+                    return array(
+                        'status'=>'error',
+                        'title'=>'No Member Selected!',
+                        'message' => 'Select the member you want to Delete!'
+                    );
+                }
+                
+            }
+            else
+            {
+                return array(
+                    'status'=>'error',
+                    'title'=>'Group is not Selected!',
+                    'message' => 'Select the group to perform this operation.'
+                );
+            }
         }
-        return array('message' => 'Friends successfully added to '.$group->name.' Group');
+        else
+            {
+                return array(
+                    'status'=>'error',
+                    'title'=>'Group is not Selected!',
+                    'message' => 'Select the group to perform this operation.'
+                );
+            }
+        
     }
 
     public function deleteGroups(Request $request)
